ES6 conversion and restructure
